+++ default.nix
@@ -19,7 +19,8 @@ let
   foldExtensions = lib.foldr lib.composeExtensions (_: _: {});
   deps = obelisk.nixpkgs.thunkSet ./dep;
   hydra-poc = import deps.hydra-poc {};
-  cardano-node = import deps.cardano-node {};  
+  cardano-node = import deps.cardano-node {};
+  cardano-wallet = import deps.cardano-wallet {};
 
   pkgs = obelisk.nixpkgs;
   livedoc-devnet-script = pkgs.runCommand "livedoc-devnet-script" { } ''
@@ -49,6 +50,7 @@ let
               cardano-node.cardano-node
               cardano-node.cardano-cli
               cardano-node.cardano-submit-api
+              cardano-wallet.cardano-wallet
               hydra-poc.hsPkgs.hydra-node.components.exes.hydra-node
               hydra-poc.hsPkgs.hydra-node.components.exes.hydra-tools
               pkgs.jq